chore: rename component to kebab-case
